Examen Final/Chinook.WebApi/Controllers/PlaylistController.cs
using Chinook.UnitOfWork;
using Microsoft.AspNetCore.Mvc;
using Chinook.Models;

namespace Chinook.WebApi.Controllers
{

    [Route("api/Playlist")]
    public class PlaylistController : BaseController
    {
        public PlaylistController(IUnitOfWork unit) : base(unit)
        {
        }

        [HttpGet]
        [Route("list/{page}/{rows}")]
        public IActionResult GetList(int page, int rows)
        {
            var startRecord = ((page - 1) * rows) + 1;
            var endRecord = page * rows;
            return Ok(_unit.Playlists.PagedList(startRecord, endRecord));
        }

        [HttpGet]
        [Route("{id:int}")]
        public IActionResult GetById(int id)
        {
            return Ok(_unit.Playlists.GetById(id));
        }

        [HttpPost]
        public IActionResult Post([FromBody] Playlist playlist)
        {
            if (ModelState.IsValid)
            {
                return Ok(_unit.Playlists.Insert(playlist));
            }
            return BadRequest(ModelState);

        }

        [HttpPut]
        public IActionResult Put([FromBody] Playlist playlist)
        {
            if (ModelState.IsValid && _unit.Playlists.Update(playlist))
            {
                return Ok(new { Message = "The Playlist is Updated" });
            }
            return BadRequest(ModelState);

        }

        [HttpDelete]
        [Route("{id}")]
        public IActionResult Delete(int? id)
        {
            if (id.HasValue && id.Value > 0)
            {
                return Ok(_unit.Playlists.Delete(new Playlist { PlaylistId = id.Value }));
            }
            return BadRequest(new { Message = "Incorrect data" });

        }

        [HttpGet]
        [Route("count")]
        public IActionResult GetCount()
        {
            return Ok(_unit.Playlists.Count());
        }

        [HttpGet]
        public IActionResult GetAllList()
        {
            return Ok(_unit.Playlists.GetList());
        }
    }
}